linted and cleaned up code
