+++ app/components/base/Main.jsx
@@ -1,21 +1,35 @@
 import React, { Component, PropTypes } from "react";
-
 import { DragDropContext } from "react-dnd";
 import HTML5Backend from "react-dnd-html5-backend";
-// import TouchBackend from "react-dnd-touch-backend";
-
 import Header from "./Header.jsx";
 import Footer from "./Footer.jsx";
-
 import { Menu, Sidebar } from "semantic-ui-react";
-
 import UnitSearchContainer from "../../containers/UnitSearchContainer.jsx";
-// import UnitDragPreview from "../Unit/UnitDragPreview.jsx";
-
-/* These imports handle the smooth transitioning between app views */
 import ReactCSSTransitionGroup from "react-addons-css-transition-group";
 import "../../resources/css/transitions.css";
+import { bindActionCreators } from "redux";
+import { connect } from "react-redux";
+import * as courseActions from "../../actions/CourseActions";
+import * as counterActions from "../../actions/CounterActions";
 
+
+/**
+ * Redux binding for main, this passes in the state as props to the main component via connect method at the bottom
+ */
+const mapStateToProps = (state) => {
+    return {
+        course: state.Course,
+        counter: state.Counter
+    };
+};
+
+/**
+ * Redux bindings for functions, passes the action creators through with dispatch built in
+ */
+const mapDispatchToProps = (dispatch) => {
+    const actionBundle = Object.assign({}, courseActions, counterActions);
+    return bindActionCreators(actionBundle, dispatch);
+};
 
 /**
  * The main layout used for all views.
@@ -152,4 +166,12 @@ Main.propTypes = {
     location: PropTypes.object
 };
 
-export default Main;
+
+/**
+ * THIS IS IMPORTANT
+ * To ensure that redux connect plays well with react drag and drop (which also uses connect), 
+ * we need to first connect the main app with drag and drop, then we take the connected dnd version and 
+ * connect it again through the redux. 
+ */
+const App = DragDropContext(HTML5Backend)(Main);
+export default connect(mapStateToProps, mapDispatchToProps)(App);
