+++ app/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render } from "react-dom";
-import { Router, browserHistory } from "react-router";
+import { Router } from "react-router";
 import { Provider } from "react-redux";
 
 import routes from "./config/routes";
