+++ app/wrappers/CourseStatisticGroupWrapper.js
@@ -1,6 +1,9 @@
 import { connect } from "react-redux";
 import CourseStatisticGroup from "../components/Course/CourseStatisticGroup.jsx";
 
+/**
+ * Course statistic group just needs the cost and credit points from the state
+ */
 const mapStateToProps = (state) => {
     return {
         cost: state.Counter.cost,
